Create infytq ass-5
Write a python program to implement the following.

insert_circle(new_circle)- inserts the new_circle as the first circle in the circle_list.
Note : circle_list is maintained as the linked list where data in each node refers to a circle. Consider circle to be an object of Circle class which has color and radius as attributes. Code for Circle class is given in the code pane.