Create Infytq ass-6
The teacher instructed the children to be ready by holding each other’s hands while standing on the stage and they have to come forward and perform when their turn comes. It is decided that Rahul would perform at the beginning and once again in the middle of the show. After Rahul’s first performance, he would join the other children, in his new position(after Gitu). After Rahul occupies the new position, Swetha would join the children at the end for delivering the vote of thanks.
Assume that there will always be odd number of children initially.